organize map related files
